Replace batch.add with ShaderProgram
Due to changes in Pyglet 2.x Graphics Module, instead of using batch.add(),
ShaderProgram must be written and used in ShaderProgram.vertex_list().
The ShaderProgram is written in the beginning of led.py, which I use to create a ShaderProgram.
Similarly, batch.add() is used in other scripts like simulator.py.
So, we could make a txt file with ShaderProgram, because writing it more than once is not efficient,
 or we could define a global ShaderProgram variable to use it every time we did batch.add().